==> 0_Analyses/0_Nf_modeling/R/2d_Model_predict_GspaceClimBioticHabitatClamping.R <==
# set working directory to this script's location:
setwd(dirname(rstudioapi::getActiveDocumentContext()$path))

# Read functions
source("../../0_Nf_modeling/GEspace.R")
source("../../0_Nf_modeling/nicheG.R")
library(ENMeval)

# Read table with estimated parameters of niche model
mles <- read.csv("../../../0_Analyses/0_Nf_modeling/ResultsClimBioticHabitat/mle_allspecies_v3.csv",header=T)


# Project models to geographic space in the native and invaded regions

for (k in 1:nrow(mles)) {
  # Select species
  species.id <- mles[k,1]
  
  # Project model into the native range
  pc.native <- stack(paste0("../../../1_Data/sdm_predictors/bioreg_occ_dist_allgrids/",paste0(species.id,"/"),
                            species.id,"_allGridsPCA.native.tif"))
  
  suit.wn.nat <- niche.G(Estck = pc.native, mu = c(mles[k,3],mles[k,4]), 
                       Sigma = matrix(c(mles[k,5], mles[k,6], mles[k,6], 
                                        mles[k,7]),ncol=2))
  
  writeRaster(suit.wn.nat,paste0("../../../0_Analyses/0_Nf_modeling/ResultsClimBioticHabitatClamping/",species.id,
                             "_suitability_map_nat.tif"), overwrite = T)
  
  # Project model into the invaded region using CLAMPING
  
  pc.invasive <- stack(paste0("../../../1_Data/sdm_predictors/bioreg_occ_dist_allgrids/",paste0(species.id,"/"),
                              species.id,"_allGridsPCA.invasive.tif"))
  
  spare.names <- names(pc.invasive)
  names(pc.invasive) <- c("PC1","PC2")
  
  j<-species.id
  f.occ.nat <- paste0("../../../0_Analyses/0_Nf_modeling/accessible_areas_ClimBioticHabitat/",j,"_native_range.csv")
  sp.occ.nat <- read.csv(f.occ.nat,header=T)[,-1]
  
  pc.invasiveCL <- clamp.vars(pc.invasive, sp.occ.nat[c(3:4)], left = NULL, right = NULL, categoricals = NULL)

  suit.wn.inv <- niche.G(Estck = pc.invasiveCL, mu = c(mles[k,3],mles[k,4]), 
                     Sigma = matrix(c(mles[k,5], mles[k,6], mles[k,6], 
                                      mles[k,7]),ncol=2))
  
  writeRaster(suit.wn.inv,paste0("../../../0_Analyses/0_Nf_modeling/ResultsClimBioticHabitatClamping/",species.id,
                             "_suitability_map_inv.tif"), overwrite = T)
}
